changed health report build ints to long because the github run id is larger than int32 max value

+++ src/Api/Dtos/HealthChecks/BuildInformation.cs
@@ -4,8 +4,8 @@ public class BuildInformation
 {
     public string? CommitSha { get; set; }
     public DateTime? CreatedOn { get; set; }
-    public int? RunAttempt { get; set; }
-    public int? RunId { get; set; }
-    public int? RunNumber { get; set; }
+    public long? RunAttempt { get; set; }
+    public long? RunId { get; set; }
+    public long? RunNumber { get; set; }
     public string? Version { get; set; }
 }
